agentic-trading-system/requirements.txt
# Agentic Trading System - Core Dependencies

# Core Framework
python>=3.11
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# Agent Orchestration
langgraph>=0.0.40
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.0.5

# LLM Providers
anthropic>=0.18.0
openai>=1.12.0

# Data Management
# Using SQLite (built-in) instead of PostgreSQL
# Using diskcache instead of Redis
diskcache>=5.6.3
pyarrow>=14.0.0
pandas>=2.1.0
numpy>=1.24.0
pyyaml>=6.0.1  # For config files

# Backtesting & Technical Analysis
backtrader>=1.9.78
ta-lib>=0.4.28
scipy>=1.11.0

# Data Sources
yfinance>=0.2.32
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-dateutil>=2.8.0

# API Integration
newsapi-python>=0.2.7
# Perplexity API (via OpenAI SDK - uses same interface)

# Async Support
aiofiles>=23.2.0

# UI (Optional)
streamlit>=1.29.0
plotly>=5.18.0
matplotlib>=3.8.0

# Monitoring & Logging
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
ipython>=8.18.0
